refactor: rename pages components and add authApi
