+++ app/profile/page.tsx
@@ -1,1 +1,1 @@
-export { default } from '@/pages/profile/ui/page'
+export { default } from '@/pages/profile/ui/Profile'
