+++ src/app/api/inctagramApi.types.ts
@@ -1,17 +1,3 @@
-export type registrationArgs = {
-  email: string
-  password: string
-  userName: string
-}
-
-export type registrationConfirmationArgs = {
-  code: string
-}
-
-export type registrationResendingArgs = {
-  email: string
-}
-
 export type User = {
   createdAt: string
   email: string
