+++ src/pages/createNewPassword/ui/newPasswordPage.module.scss
