+++ src/pages/forgotPasswordPage/ui/ForgotPasswordPage.tsx
@@ -2,9 +2,9 @@
 
 import { ForgotPasswordForm } from '@/shared/ui/forms/forgotPassword'
 
-import s from './page.module.scss'
+import s from './forgotPasswordPage.module.scss'
 
-export default function Page() {
+export default function ForgotPasswordPage() {
   return (
     <div>
       <main>
