+++ src/pages/forgotPasswordPage/ui/forgotPasswordPage.module.scss
