+++ src/pages/profile/ui/Profile.tsx
@@ -1,8 +1,11 @@
 'use client'
 
+import { Sidebar } from '@/shared/ui/sidebar'
+
 export default function Home() {
   return (
     <div>
+      <Sidebar />
       <main>Profile Page</main>
     </div>
   )
